simulacra-web/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.2
python-multipart>=0.0.7
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
matplotlib>=3.7.1
openai>=1.0.0,<2.0.0
openai-agents
